refactor: use Suspense for initial page loading
