+++ __mocks__/firebase/auth.ts
@@ -4,5 +4,13 @@ export const getAuth = jest.fn(() => ({
   },
 }))
 export const signInAnonymously = jest.fn()
-export const onAuthStateChanged = jest.fn()
-export const onIdTokenChanged = jest.fn()
+export const onAuthStateChanged = jest.fn((_, nextOrObserver) => {
+  nextOrObserver({})
+  return jest.fn()
+})
+export const onIdTokenChanged = jest.fn((_, nextOrObserver) => {
+  nextOrObserver({
+    getIdToken: jest.fn(() => Promise.resolve('id')),
+  })
+  return jest.fn()
+})
