+++ src/components/shared/app/index.ts
@@ -1,1 +1,2 @@
 export * from './BeforeAppMount'
+export * from './Subscription'
